+++ naampy/__init__.py
@@ -1,5 +1,0 @@
-"""naampy: infer gender from Indian first names."""
-
-from .in_rolls_fn import InRollsFnData, in_rolls_fn_gender, predict_fn_gender
-
-__all__ = ["InRollsFnData", "in_rolls_fn_gender", "predict_fn_gender"]